v2.distToSq(v2) bertipe data int, bukan double
Hasil jumlah kuadrat bilangan-bilangan bulat adalah bilangan bulat juga.

+++ src/maze/v2.java
@@ -18,7 +18,7 @@ public class v2 {
 		return (target.x == x) && (target.y == y);
 	}
 
-	public double distToSq(v2 target) {
+	public int distToSq(v2 target) {
 		return (target.x - x) * (target.x - x) +
 				(target.y - y) * (target.y - y);
 	}
